Implement functionality for a variable number of packages.
In helper.py, package reading from .csv is improved. The number of packages is now taken from the number of rows containing package data. Previously it was assumed that exactly 40 packages would be delivered.

The hash table is now set up with more buckets. It follows a standard of a prime number at least 1.3 times the number of objects to be stored.